+++ src/main/java/io/github/doocs/im/model/response/ImportMsgResultItem.java
@@ -8,7 +8,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
  */
 public class ImportMsgResultItem {
     @JsonProperty("MsgSeq")
-    private Integer msgSeq;
+    private Long msgSeq;
 
     @JsonProperty("MsgTime")
     private Integer msgTime;
@@ -16,11 +16,11 @@ public class ImportMsgResultItem {
     @JsonProperty("Result")
     private Integer result;
 
-    public Integer getMsgSeq() {
+    public Long getMsgSeq() {
         return msgSeq;
     }
 
-    public void setMsgSeq(Integer msgSeq) {
+    public void setMsgSeq(Long msgSeq) {
         this.msgSeq = msgSeq;
     }
