+++ src/main/java/io/github/doocs/im/model/response/SendGroupMsgResult.java
@@ -11,7 +11,7 @@ public class SendGroupMsgResult extends GenericResult {
     private Integer msgTime;
 
     @JsonProperty("msgSeq")
-    private Integer msgSeq;
+    private Long msgSeq;
 
     public Integer getMsgTime() {
         return msgTime;
@@ -21,11 +21,11 @@ public class SendGroupMsgResult extends GenericResult {
         this.msgTime = msgTime;
     }
 
-    public Integer getMsgSeq() {
+    public Long getMsgSeq() {
         return msgSeq;
     }
 
-    public void setMsgSeq(Integer msgSeq) {
+    public void setMsgSeq(Long msgSeq) {
         this.msgSeq = msgSeq;
     }
